fixed an issue on virtual-list.js
Now it checks if the vl.listBlock element is visible before it does the resizing- and rendering-call. Like this, the virtual list won't be overwritten with wrong content when it's not visible. This happens when you have a routing with more than 1 level because just the current back-page keeps visible.