Fix Problem with {{@index}} variable in virtual List
Fix a Problem with template-line-index {{@index}} variable in virtual List. 
In this line the used index-variable for a compiled template is in form 
{index: index-number} expected. 
In version 2.x of framework7 there is only a number Variable. 
Inside a compiled template the data access is done by
  c((data && data.index), ctx_1)
so the compiled template needs a Object with variable index.