fixing the background color property for action sheet buttons
The background color wasn't being applied correctly due to the class name being incorrect. I changed the class name to include -color- to fix the issue. I also noticed that the non popover option for the action sheet was using the color variable instead of bg variable. I changed it to use the bg variable just like the popover version of the action sheet does.