Implement clearDomCache method in the Router
Factor out the Dom cache clearing from clearPreviousHistory(), and export it as a Router method.

This facilitates a use case like chaning language preference, where you can refreshPage() to reload the current page in the current language, and then call clearDomCache() to ensure that pages in the history will be reloaded in the current language if they are revisited.