FIX: [react] pass props from route options for initial route
getRouterInitialComponent() function sets up components for the default route.
Currently, props configured on the route options are not passed when creating
the initial component, so are not available on first render. (props are
available only on subsequent navigation). This change fixes the issue by copying
props from route options when setting up initial component