z-index of popup-behind
Currently, when several popups are stacked, the backdrop is not visible for the closest one.
By adding a z-index to the behind popup smaller than the one of the backdrop, we make the backdrop visible.

This solution has two issues:
1/ There is no "animation" when the backdrop of the stacked popup appears
2/ Since all the popups are using the same backdrop element, it's better to have all popups behaving the same (= all or none popup should have a backdrop, and all or none popup should have a click-to-close backdrop)

A better solution would be to generate a dedicated backdrop for each popup, with a z-index dynamically computed to be correctly sandwiched between the popups (or by having all the same z-index bu by being inserted in the right order). But it means a lot of recode in comparison to this one-line solution.